+++ Venmo/src/test/java/com/braintreepayments/api/VenmoRequestUnitTest.java
@@ -49,6 +49,16 @@ public class VenmoRequestUnitTest {
     }
 
     @Test
+    public void getIsFinalAmountAsString_returnsStringEquivalent() {
+        VenmoRequest request = new VenmoRequest(VenmoPaymentMethodUsage.MULTI_USE);
+        request.setIsFinalAmount(true);
+        assertEquals("true", request.getIsFinalAmountAsString());
+
+        request.setIsFinalAmount(false);
+        assertEquals("false", request.getIsFinalAmountAsString());
+    }
+
+    @Test
     public void parcelsCorrectly() {
         VenmoRequest request = new VenmoRequest(VenmoPaymentMethodUsage.MULTI_USE);
         request.setDisplayName("venmo-user");
@@ -61,6 +71,7 @@ public class VenmoRequestUnitTest {
         request.setDiscountAmount("2.00");
         request.setShippingAmount("1.00");
         request.setTotalAmount("10.00");
+        request.setIsFinalAmount(true);
 
         ArrayList<VenmoLineItem> lineItems = new ArrayList<>();
         lineItems.add(new VenmoLineItem(VenmoLineItem.KIND_DEBIT, "An Item", 1, "10.00"));
@@ -84,5 +95,6 @@ public class VenmoRequestUnitTest {
         assertEquals("10.00", result.getTotalAmount());
         assertEquals(1, result.getLineItems().size());
         assertEquals("An Item", result.getLineItems().get(0).getName());
+        assertTrue(result.getIsFinalAmount());
     }
 }